Fixed bugs in the DHT11 sensor driver

+++ Src/GPIO/Input/DHT11Sensor.cs
@@ -3,7 +3,7 @@ using Core.Device;
 
 namespace GPIO
 {
-	public class DHT11Sensor: RemotlyAccessableDeviceBase, IInputMeter<float>
+	public class DHT11Sensor: RemotlyAccessableDeviceBase, IInputMeter<int>
 	{
 		private IDHT11 m_dht;
 		private DHT11ValueType m_type;
@@ -21,7 +21,7 @@ namespace GPIO
 
 		}
 
-		public float Value { get { return m_type == DHT11ValueType.Temperature ? m_dht.Temperature : m_dht.Humidity; } }
+		public int Value { get { return m_type == DHT11ValueType.Temperature ? m_dht.Temperature : m_dht.Humidity; } }
 
 		#region IRemotlyAccessable implementation
 		public override byte[] RemoteRequest (string methodName, byte[] rawData, IRPCManager<System.Net.IPEndPoint> mgr)
@@ -29,7 +29,7 @@ namespace GPIO
 			if (IsBaseMethod (methodName)) {
 				return ExecuteStandardDeviceMethod (methodName, rawData, mgr);
 			} else if (methodName == RemoteInputMeter.GET_VALUE_FUNCTION_NAME) {
-				return mgr.RPCReply<float> (Guid, methodName, Value);
+				return mgr.RPCReply<int> (Guid, methodName, Value);
 			} else
 				throw new NotImplementedException ("Method name: " + methodName + " is not implemented for Distance meter.");
