+++ Src/GPIO/Input/IDHT11.cs
@@ -12,27 +12,27 @@ namespace GPIO
 		/// Temperature in C
 		/// </summary>
 		/// <value>The temperature.</value>
-		float Temperature { get; }
+		int Temperature { get; }
 
 		/// <summary>
 		/// Relative humidity in percentage
 		/// </summary>
 		/// <value>The humidity.</value>
-		float Humidity { get; }
+		int Humidity { get; }
 
 		/// <summary>
 		/// Returns an analog sensor for measuring the humidity of me.
 		/// </summary>
 		/// <returns>The humidity sensor.</returns>
 		/// <param name="id">Identifier.</param>
-		IInputMeter<float> GetHumiditySensor(string id);
+		IInputMeter<int> GetHumiditySensor(string id);
 
 		/// <summary>
 		/// Returns an analog sensor for measuringing my temperature.
 		/// </summary>
 		/// <returns>The temperature sensor.</returns>
 		/// <param name="id">Identifier.</param>
-		IInputMeter<float> GetTemperatureSensor(string id);
+		IInputMeter<int> GetTemperatureSensor(string id);
 
 	}
 }
